Bump Chart.js v2.9.3 to v3.2.1
Patch done based on security alert CVE-2020-7746

Prototype pollution in chart.js - https://github.com/advisories/GHSA-h68q-55jf-x68w

This affects the package chart.js before 2.9.4. The options parameter is not properly sanitized when it is processed. When the options are processed, the existing options (or the defaults options) are deeply merged with provided options. However, during this operation, the keys of the object being set are not checked, leading to a prototype pollution.